+++ Cargo.toml
@@ -40,6 +40,7 @@ members = [
   "oak_functions_sdk/tests/lookup_module",
   "oak_functions_sdk/tests/testing_module",
   "oak_functions_service",
+  "oak_functions_enclave_service",
   "oak_functions_test_utils",
   "oak_hello_world_linux_init",
   "oak_launcher_utils",
@@ -98,6 +99,7 @@ oak_functions_launcher = { path = "./oak_functions_launcher" }
 oak_functions_lookup = { path = "./oak_functions/lookup" }
 oak_functions_sdk = { path = "./oak_functions_sdk" }
 oak_functions_service = { path = "./oak_functions_service", default-features = false }
+oak_functions_enclave_service = { path = "./oak_functions_enclave_service", default-features = false }
 oak_functions_test_utils = { path = "./oak_functions_test_utils" }
 oak_functions_wasm = { path = "./oak_functions/wasm" }
 oak_grpc_utils = { path = "./oak_grpc_utils" }
