+++ enclave_apps/oak_functions_enclave_app/Cargo.toml
@@ -8,13 +8,15 @@ license = "Apache-2.0"
 [features]
 default = ["deny_sensitive_logging"]
 # Disable sensitive logging.
-deny_sensitive_logging = ["oak_functions_service/deny_sensitive_logging"]
+deny_sensitive_logging = [
+  "oak_functions_enclave_service/deny_sensitive_logging"
+]
 # Feature allow_sensitive_logging is not actually used in the code. It is only used as a
 # required feature to differentiate between the two binaries.
 allow_sensitive_logging = []
 
 [dependencies]
-oak_functions_service = { path = "../../oak_functions_service", default-features = false }
+oak_functions_enclave_service = { path = "../../oak_functions_enclave_service", default-features = false }
 log = "*"
 micro_rpc = { workspace = true }
 oak_attestation = { workspace = true }
